\name{testgenotypes}
\docType{data}
\alias{testgenotypes}
\title{Rubus Genotype Data for Learning polysat}
\description{
  \code{genambig} object containing alleles of 20 \emph{Rubus}
  samples at three microsatellite loci.
}
\usage{data(testgenotypes)}
\format{A \code{genambig} object with data in the \code{Genotypes},
  \code{PopInfo}, \code{PopNames}, and \code{Usatnts} slots.  This is
  saved as a .RData file.  Population identities are used here to
  indicate two different species.}
\source{Clark, L. V. and Jasieniuk, M. (2012)  Spontaneous
  hybrids between native and exotic \emph{Rubus} in the Western United
  States produce offspring both by apomixis and by sexual
  recombination. \emph{Heredity} \bold{109}, 320--328.  Data available
  at: \url{http://dx.doi.org/10.5061/dryad.m466f}}
\seealso{\code{\link{FCRinfo}}, \code{\link{simgen}},
  \code{\linkS4class{genambig}}}
\keyword{datasets}