\name{FCRinfo}
\docType{data}
\alias{FCRinfo}
\title{Additional Data on Rubus Samples}
\description{
  For 20 Rubus samples, contains numbers indicating species identity, as
  well as colors and symbols to use for plotting data.
}
\usage{data(FCRinfo)}
\format{Data frame.  \code{FCRinfo$Species} contains integers indicating the
  species of each sample.  \code{FCRinfo$Plot.color} contains character
  strings of the colors that each sample should be plotted as.
  \code{FCR.info$Plot.symbol} contains integers to be passed to
  \code{pch} to designate the symbol used to represent each individual.}
\source{Clark and Jasieniuk, unpublished data}
\seealso{\code{\link{testgenotypes}}}
\keyword{datasets}