\name{FCRinfo}
\docType{data}
\alias{FCRinfo}
\title{Additional Data on Rubus Samples}
\description{
  For 20 \emph{Rubus} samples, contains colors and symbols to use for
  plotting data.
}
\usage{data(FCRinfo)}
\format{Data frame.  \code{FCRinfo$Plot.color} contains character
  strings of the colors to be used to represent species groups.
  \code{FCR.info$Plot.symbol} contains integers to be passed to
  \code{pch} to designate the symbol used to represent each individual.
These reflect chloroplast haplotypes.}
\source{Clark and Jasieniuk, unpublished data}
\seealso{\code{\link{testgenotypes}}}
\keyword{datasets}