\name{testgenotypes}
\docType{data}
\alias{testgenotypes}
\title{Rubus Genotype Data for Learning polysat}
\description{
  \code{genambig} object containing alleles of 20 \emph{Rubus}
  samples at three microsatellite loci.
}
\usage{data(testgenotypes)}
\format{A \code{genambig} object with data in the \code{Genotypes},
  \code{PopInfo}, \code{PopNames}, and \code{Usatnts} slots.  This is
  saved as a .RData file.  Population identities are used here to
  indicate two different species.}
\source{Clark and Jasieniuk, unpublished data}
\seealso{\code{\link{FCRinfo}}, \code{\link{simgen}},
  \code{\linkS4class{genambig}}}
\keyword{datasets}