\name{testgenotypes}
\docType{data}
\alias{testgenotypes}
\title{Rubus Genotype Data for Learning polysat}
\description{
  Contains alleles of 20 Rubus samples at three microsatellite loci.
}
\usage{data(testgenotypes)}
\format{A two-dimensional list of integer vectors.  Samples are the
  first dimension of the list and loci the second dimension.  Each
  vector contains all alleles for a given sample and locus.}
\source{Clark and Jasieniuk, unpublished data}
\seealso{\code{\link{FCRinfo}}}
\keyword{datasets}